Move video controls and settings into their own component files
