+++ features/players/components/video-controls/index.ts
@@ -1,0 +1,1 @@
+export { VideoControls } from "./VideoControls";
