+++ features/players/components/video-settings/VideoSettings.tsx
@@ -2,8 +2,8 @@ import { useKeyboardNavigation } from "hooks/useKeyboardMenuNavigation";
 import { useMenuCloseEvents } from "hooks/useMenuCloseEvents";
 import { Player } from "features/players";
 import { useState } from "react";
-import { QualitySettingsMenu } from "./video-settings/QualitySettingsMenu";
-import { PlaybackSpeedSettingsMenu } from "./video-settings/PlaybackSpeedSettingsMenu";
+import { QualitySettingsMenu } from "./QualitySettingsMenu";
+import { PlaybackSpeedSettingsMenu } from "./PlaybackSpeedSettingsMenu";
 import * as React from "react";
 import styles from "features/players/components/styles/VideoSettings.module.css";
 import ArrowBackIcon from "icons/ArrowBackIcon";
