+++ features/players/components/video-settings/index.ts
@@ -1,0 +1,1 @@
+export { VideoSettings } from "./VideoSettings";
