+++ features/players/index.ts
@@ -1,9 +1,9 @@
 export * from "./components/twitch-player";
 export * from "./components/youtube-player";
 export * from "./components/VideoContainer";
-export * from "./components/VideoControls";
+export * from "./components/video-controls";
 export * from "./components/VideoPlayer";
-export * from "./components/VideoSettings";
+export * from "./components/video-settings";
 export * from "./api/player";
 export * from "./api/useTwitchPlayer";
 export * from "./api/useYouTubeIframe";
